% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{average_test_margin}
\alias{average_test_margin}
\title{Average test margin}
\usage{
average_test_margin(alpha, n = 7)
}
\arguments{
\item{alpha}{Interaction factor}

\item{n}{Number of items}
}
\description{
Returns the average test margin
}
