% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behav-data.R
\docType{data}
\name{ciranka_data}
\alias{ciranka_data}
\title{Behavioral data from humans in Ciranka et al. (2022)}
\format{
\subsection{\code{ciranka_data}}{

\describe{
\item{symbolic_distance}{Symbolic distance between the target and the
distractor}
\item{subject}{The subject}
\item{acc}{The proportion of correct decisions.}
}
}
}
\source{
\link{https://doi.org/10.1371/journal.pcbi.1004523.s002}
}
\usage{
ciranka_data
}
\description{
Data from Ciranka et al. (2022) performing transitive inference.
This is organised according to symbolic distance as this is the only data
we plot in the figures.
}
\seealso{
Other {behavioral datasets}: 
\code{\link{jensen_data_humans}},
\code{\link{jensen_data_monkeys}}
}
\concept{{behavioral datasets}}
\keyword{datasets}
