% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_ti.R
\docType{data}
\name{depth_ti}
\alias{depth_ti}
\title{TI behavior for different depths}
\format{
\subsection{\code{depth_ti}}{

\describe{
\item{model_seed}{Random seed of the model initialization}
\item{depth}{Number of hidden layers}
\item{j,k}{Indices of the items being compared}
\item{margin} {Network output}
}
}
}
\usage{
depth_ti
}
\description{
TI behavior for different depths
}
\seealso{
Other {simulations}: 
\code{\link{crossentropy_ti_rank_rep}},
\code{\link{crossentropy_ti}},
\code{\link{depth_ti_rank_rep}},
\code{\link{n_ti_rank_rep}},
\code{\link{n_ti}},
\code{\link{nonlinearity_ti_rank_rep}},
\code{\link{nonlinearity_ti}},
\code{\link{similarity}},
\code{\link{tanh_ti_rank_rep}},
\code{\link{tanh_ti}},
\code{\link{ti_scaling_rank_rep}},
\code{\link{ti_scaling}},
\code{\link{time_traj_ti_rank_rep}},
\code{\link{time_traj_ti}}
}
\concept{{simulations}}
\keyword{datasets}
