% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{gd_adjacent_margin}
\alias{gd_adjacent_margin}
\title{Adjacent margins over the course of gradient descent}
\usage{
gd_adjacent_margin(j, t = Inf, n = 7, alpha = 0)
}
\arguments{
\item{j}{Integer between 1 and n-1.}

\item{t}{Time over gradient descent. Default is \code{Inf} which corresponds to
the end of training.}

\item{n}{Number of items. Default is 7.}

\item{alpha}{Interaction factor (between 0 and 1).}
}
\value{
The margin between item j and j+1.
}
\description{
Returns the margin for adjacent items over the course of gradient descent.
}
