% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{gd_margin}
\alias{gd_margin}
\title{Margin over the course of gradient descent.}
\usage{
gd_margin(j, k, t = Inf, n = 7, alpha = 0)
}
\arguments{
\item{j, k}{Integers between 1 and n.}

\item{t}{Time over gradient descent. Default is \code{Inf} which corresponds to
the end of training.}

\item{n}{Number of items. Default is 7.}

\item{alpha}{Interaction factor (between 0 and 1).}
}
\value{
The margin between item j and item k.
}
\description{
Returns the margin given in Lemma S1.4.
}
