% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{gd_rank}
\alias{gd_rank}
\title{Rank representation over the course of gradient descent}
\usage{
gd_rank(j, t = Inf, n = 7, alpha = 0)
}
\arguments{
\item{j}{Integer between 1 and n. Index of the item for which the rank is
returned.}

\item{t}{Time over gradient descent. Default is \code{Inf} which corresponds to
the end of training.}

\item{n}{Number of items. Default is 7.}

\item{alpha}{Interaction factor (between 0 and 1).}
}
\value{
The rank of the j-th item.
}
\description{
Returns the rank representation given in Lemma S1.4.
}
