% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffs.R
\name{get_alphas}
\alias{get_alphas}
\title{Interaction factors for deep networks}
\usage{
get_alphas(rho, l, lambd = 0, type = "hidden_layer")
}
\arguments{
\item{rho}{Parameter of nonlinearity (between 0 and 2). \code{rho=1} corresponds
to ReLU}

\item{l}{Maximal depth}

\item{lambd}{Proportion of additive nodes. Default is 0.}

\item{type}{If \code{hidden_layer}, computes the interaction factor of the hidden
layers, if \code{ntk}, computes the interaction factor of the NTK.}
}
\description{
Returns interaction factors of deep networks.
}
