% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeffs.R
\name{get_kappas_ntk}
\alias{get_kappas_ntk}
\title{Neural tangent kernel for deep networks}
\usage{
get_kappas_ntk(rho, l, lambd = 0)
}
\arguments{
\item{rho}{Parameter of nonlinearity (between 0 and 2). \code{rho=1} corresponds
to ReLU}

\item{l}{Maximal depth}

\item{lambd}{Proportion of additive nodes. Default is 0.}
}
\description{
Returns the interaction factors of deep networks' NTK.
}
