% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_rep.R
\name{get_rank_rep}
\alias{get_rank_rep}
\title{Compute rank representability}
\usage{
get_rank_rep(dat)
}
\arguments{
\item{dat}{Dataframe with \code{j} and \code{k} columns.}
}
\value{
A scalar error.
}
\description{
The rank representability is defined as the mean squared residual error of
approximating the model behavior by a rank representation.
}
