% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behav-data.R
\docType{data}
\name{jensen_data_humans}
\alias{jensen_data_humans}
\title{Behavioral data from humans in Jensen et al. (2015)}
\format{
\subsection{\code{jensen_data_humans}}{

\describe{
\item{symbolic_distance}{Symbolic distance between the target and the
distractor}
\item{subject}{The subject}
\item{acc}{The proportion of correct decisions.}
}
}
}
\source{
\link{https://doi.org/10.1371/journal.pcbi.1004523.s002}
}
\usage{
jensen_data_humans
}
\description{
Data from Jensen et al. (2015) on humans performing transitive inference.
This is organised according to symbolic distance as this is the only data
we plot in the figures.
}
\seealso{
Other {behavioral datasets}: 
\code{\link{ciranka_data}},
\code{\link{jensen_data_monkeys}}
}
\concept{{behavioral datasets}}
\keyword{datasets}
