% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behav-data.R
\docType{data}
\name{jensen_data_monkeys}
\alias{jensen_data_monkeys}
\title{Behavioral data from monkeys in Jensen et al. (2015)}
\format{
\subsection{\code{jensen_data_monkeys}}{

\describe{
\item{session}{The session}
\item{target}{The index of the target (between 1 and 6).}
\item{distractor}{The index of the distractor (between 2 and 7).}
\item{acc}{The proportion of correct decisions.}
}
}
}
\source{
\link{https://doi.org/10.1371/journal.pcbi.1004523.s002}
}
\usage{
jensen_data_monkeys
}
\description{
Data from Jensen et al. (2015) on monkeys performing transitive inference.
}
\seealso{
Other {behavioral datasets}: 
\code{\link{ciranka_data}},
\code{\link{jensen_data_humans}}
}
\concept{{behavioral datasets}}
\keyword{datasets}
