% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_ti.R
\docType{data}
\name{n_ti}
\alias{n_ti}
\title{TI behavior for different numbers of items}
\format{
\subsection{\code{n_ti}}{

\describe{
\item{n}{Number of items}
\item{model_seed}{Random seed of the model initialization}
\item{j,k}{Indices of the items being compared}
\item{margin} {Network output}
}
}
}
\usage{
n_ti
}
\description{
TI behavior for different numbers of items
}
\seealso{
Other {simulations}: 
\code{\link{crossentropy_ti_rank_rep}},
\code{\link{crossentropy_ti}},
\code{\link{depth_ti_rank_rep}},
\code{\link{depth_ti}},
\code{\link{n_ti_rank_rep}},
\code{\link{nonlinearity_ti_rank_rep}},
\code{\link{nonlinearity_ti}},
\code{\link{similarity}},
\code{\link{tanh_ti_rank_rep}},
\code{\link{tanh_ti}},
\code{\link{ti_scaling_rank_rep}},
\code{\link{ti_scaling}},
\code{\link{time_traj_ti_rank_rep}},
\code{\link{time_traj_ti}}
}
\concept{{simulations}}
\keyword{datasets}
