% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonlinearity_ti.R
\docType{data}
\name{nonlinearity_ti}
\alias{nonlinearity_ti}
\title{TI behavior for different piecewise linear activation functions}
\format{
\subsection{\code{nonlinearity_ti}}{

\describe{
\item{rho}{Scalar parameterizing the activation function (see (59))}
\item{model_seed}{Random seed of the model initialization}
\item{j,k}{Indices of the items being compared}
\item{margin} {Network output}
}
}
}
\usage{
nonlinearity_ti
}
\description{
TI behavior for different piecewise linear activation functions
}
\seealso{
Other {simulations}: 
\code{\link{crossentropy_ti_rank_rep}},
\code{\link{crossentropy_ti}},
\code{\link{depth_ti_rank_rep}},
\code{\link{depth_ti}},
\code{\link{n_ti_rank_rep}},
\code{\link{n_ti}},
\code{\link{nonlinearity_ti_rank_rep}},
\code{\link{similarity}},
\code{\link{tanh_ti_rank_rep}},
\code{\link{tanh_ti}},
\code{\link{ti_scaling_rank_rep}},
\code{\link{ti_scaling}},
\code{\link{time_traj_ti_rank_rep}},
\code{\link{time_traj_ti}}
}
\concept{{simulations}}
\keyword{datasets}
