% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins.R
\name{reg_margin}
\alias{reg_margin}
\title{Margin for regularized regression}
\usage{
reg_margin(j, k, n = 7, alpha = 0, c = Inf, lambda = NULL)
}
\arguments{
\item{j, k}{Integers between 1 and n.}

\item{n}{Number of items. Default is 7.}

\item{alpha}{Conjunctivity factor (between 0 and 1).}

\item{c}{Inverse regularization coefficient Default is \code{Inf} which corresponds to
infinitesimal regularization.}
}
\value{
The margin between item j and item k.
}
\description{
Returns the margin given in Lemma S1.8.
}
