% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\docType{data}
\name{similarity}
\alias{similarity}
\title{Similarity (Experiment 1)}
\format{
\subsection{\code{similarity}}{

\describe{
\item{hdims}{Number of hidden units}
\item{model_seed}{Seed of the network simulation}
\item{similarity}{Similarity}
\item{type}{Type of the item pair (same, overlapping, distinct)}
}
}
}
\usage{
similarity
}
\description{
This simulation presents the similarity of 20 samples of slim (1000 hidden units)
and wide (10,000 hidden units) neural networks with one hidden layer.
}
\details{
Each row contains of one item pair.
}
\seealso{
Other {simulations}: 
\code{\link{crossentropy_ti_rank_rep}},
\code{\link{crossentropy_ti}},
\code{\link{depth_ti_rank_rep}},
\code{\link{depth_ti}},
\code{\link{n_ti_rank_rep}},
\code{\link{n_ti}},
\code{\link{nonlinearity_ti_rank_rep}},
\code{\link{nonlinearity_ti}},
\code{\link{tanh_ti_rank_rep}},
\code{\link{tanh_ti}},
\code{\link{ti_scaling_rank_rep}},
\code{\link{ti_scaling}},
\code{\link{time_traj_ti_rank_rep}},
\code{\link{time_traj_ti}}
}
\concept{{simulations}}
\keyword{datasets}
