% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tanh_ti.R
\docType{data}
\name{tanh_ti_rank_rep}
\alias{tanh_ti_rank_rep}
\title{Rank representability for tanh activation function}
\format{
\subsection{\code{tanh_ti_rank_rep}}{

\describe{
\item{model_seed}{Random seed of the model initialization}
\item{rank_rep} {Rank representability}
}
}
}
\usage{
tanh_ti_rank_rep
}
\description{
Rank representability for tanh activation function
}
\seealso{
Other {simulations}: 
\code{\link{crossentropy_ti_rank_rep}},
\code{\link{crossentropy_ti}},
\code{\link{depth_ti_rank_rep}},
\code{\link{depth_ti}},
\code{\link{n_ti_rank_rep}},
\code{\link{n_ti}},
\code{\link{nonlinearity_ti_rank_rep}},
\code{\link{nonlinearity_ti}},
\code{\link{similarity}},
\code{\link{tanh_ti}},
\code{\link{ti_scaling_rank_rep}},
\code{\link{ti_scaling}},
\code{\link{time_traj_ti_rank_rep}},
\code{\link{time_traj_ti}}
}
\concept{{simulations}}
\keyword{datasets}
