% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_princti_void}
\alias{theme_princti_void}
\title{Void theme}
\usage{
theme_princti_void()
}
\description{
This theme is used for the manuscript's figures if they should rely on
\code{\link[ggplot2:ggtheme]{ggplot2::theme_void()}}
}
