% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ti_scaling.R
\docType{data}
\name{ti_scaling}
\alias{ti_scaling}
\title{TI performance of neural networks at different scales}
\format{
\subsection{\code{ti_scaling}}{

\describe{
\item{scaling}{Constant by which the output was scaled}
\item{model_seed}{Random seed of the model initialization}
\item{j,k}{Indices of the items being compared}
\item{margin} {Network output}
}
}
}
\usage{
ti_scaling
}
\description{
Predictions of the network at different output scales.
}
\seealso{
Other {simulations}: 
\code{\link{crossentropy_ti_rank_rep}},
\code{\link{crossentropy_ti}},
\code{\link{depth_ti_rank_rep}},
\code{\link{depth_ti}},
\code{\link{n_ti_rank_rep}},
\code{\link{n_ti}},
\code{\link{nonlinearity_ti_rank_rep}},
\code{\link{nonlinearity_ti}},
\code{\link{similarity}},
\code{\link{tanh_ti_rank_rep}},
\code{\link{tanh_ti}},
\code{\link{ti_scaling_rank_rep}},
\code{\link{time_traj_ti_rank_rep}},
\code{\link{time_traj_ti}}
}
\concept{{simulations}}
\keyword{datasets}
