% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_traj_ti.R
\docType{data}
\name{time_traj_ti_lr}
\alias{time_traj_ti_lr}
\title{Learning rates of Exp. 2}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 3 columns.
}
\usage{
time_traj_ti_lr
}
\description{
Learning rates in Exp. 2
}
\keyword{datasets}
