% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_traj_ti.R
\docType{data}
\name{time_traj_ti_rank_rep}
\alias{time_traj_ti_rank_rep}
\title{Time trajectory of TI rank representability (Exp. 2)}
\format{
\subsection{\code{time_traj_ti}}{

\describe{
\item{scaling}{Constant by which the output was scaled}
\item{model_seed}{Random seed of the model initialization}
\item{rank_rep} {Rank representability}
\item{epoch}{Training epoch}
}
}
}
\usage{
time_traj_ti_rank_rep
}
\description{
Rank representability of the network over the duration of gradient descent.
}
\seealso{
Other {simulations}: 
\code{\link{crossentropy_ti_rank_rep}},
\code{\link{crossentropy_ti}},
\code{\link{depth_ti_rank_rep}},
\code{\link{depth_ti}},
\code{\link{n_ti_rank_rep}},
\code{\link{n_ti}},
\code{\link{nonlinearity_ti_rank_rep}},
\code{\link{nonlinearity_ti}},
\code{\link{similarity}},
\code{\link{tanh_ti_rank_rep}},
\code{\link{tanh_ti}},
\code{\link{ti_scaling_rank_rep}},
\code{\link{ti_scaling}},
\code{\link{time_traj_ti}}
}
\concept{{simulations}}
\keyword{datasets}
